warning solved

+++ misc/sizeof.c
@@ -5,7 +5,7 @@
 int main(int argc, char **argv)
 {
 	//affiche: void *      int     long
-	printf("%i %i %i",sizeof(void *), sizeof(int),sizeof(long));
+	printf("%i %i %i",(int)sizeof(void *), (int)sizeof(int),(int)sizeof(long));
 	return 0;
 }
